Add FilterMethodsTrait.
